requirements.txt
# Requirements will be added as we go

google-generativeai
python-dotenv
requests
pyttsx3
chardet
pymupdf
customtkinter
PIL
pyaudio
SpeechRecognition
deep-translator
google-search-results
pyautogui 
pygetwindow 
pywinauto



